--- src/main/java/com/moviestore/repository/MovieRepository.java
@@ -1,12 +1,0 @@
-package com.moviestore.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.moviestore.model.Movie;
-
-
-public interface MovieRepository extends JpaRepository<Movie, Long>{
-
-	
-	
-}
