+++ src/main/java/com/moviestore/repository/UserMovieRepository.java
@@ -1,6 +1,6 @@
 package com.moviestore.repository;
 
-import java.util.Collection;
+import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
@@ -8,6 +8,6 @@ import com.moviestore.model.UserMovie;
 
 public interface UserMovieRepository extends JpaRepository<UserMovie, Long>{
 	
-	Collection<UserMovie> findByReturned(boolean returned);
+	List<UserMovie> findByReturned(boolean returned);
 	
 }
